Move `APIError` to `errors.js`
Suggested here:
https://github.com/hypothesis/lms/pull/3178#pullrequestreview-768896614.
